fit: split up configuration for signed vs non-signed
Make if possible via the SIGN compiler flag decide whether we enable
signature verification or not. When enabled, keys and certificates will
be created. However, that operation is unnecessary when running and
testing plain (unsigned) configurations and therefore we should not run
those targets when the SIGN flag isn't in use.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>