Makefile: Don't delete downloads on clean
Don't delete the buildroot dl (download) dir on clean.
For distclean do move it out of the way but keep it.

On 2021/2/28 everything related to *.gnu.org was hanging after http
connect.  buildroot would eventual give up and try mirrors but that took
20 mins per file.

The situation is fixed but keeping the dl dir seems appropriate.

Signed-off-by: Bill Mills <bill.mills@linaro.org>