Makefile: don't mix building sub projects
Don't do parallel targets at the top level but do in the sub-components.
(Don't mix compiling linux with qemu, etc)
I measured w/ and w/o the .NOTPARALLEL and the time is the same for all
and for a null rebuild.  The output logs are just easier to understand.

QEMU was complaining that it could not run in parallel and to add the +.
It did cut down the build time.

Signed-off-by: Bill Mills <bill.mills@linaro.org>