QEMU: use built in tftp server, clean up dhcp in u-boot
Using the built-in tftp server makes things simpler to setup and does not
conflict with any tftp setup of user's host machine.

(I use tftpd-hpa on my machine which does a chroot into /srv/tftp so
symlinks out of that dir do not work.  This project's current Makefile
uses and assumes use of symlinks a lot so is not compatible with tftpd-hpa.
tftpd-hpa is the debian preferred server for pxie boot scenarios.)

In u-boot var scripts, handle dhcp in do_dhcp sub script.   It allows use
of a static tftp server ip but does not enforce it.  It also cleans up
warnings from attempting to autoload the bootfile.

Signed-off-by: Bill Mills <bill.mills@linaro.org>